+++ service/controller/resource/rbac/resource.go
@@ -3,7 +3,7 @@ package rbac
 import (
 	"reflect"
 
-	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/rbac/v1beta1"