+++ service/controller/resource/wrapper/monitoringdisabledresource/resource.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"github.com/giantswarm/operatorkit/v4/pkg/resource"
+	"github.com/giantswarm/operatorkit/v7/pkg/resource"
 
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )